extracts retrofitCall() to an interface and implements in the repository
